requirements.txt
absl-py==1.4.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
google-auth-oauthlib==1.0.0
google-auth==2.17.3
grpcio==1.54.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
pip==21.3.1
protobuf==4.22.4
pyasn1-modules==0.3.0
pyasn1==0.5.0
requests-oauthlib==1.3.1
requests==2.30.0
rsa==4.9
scipy==1.10.1
setuptools==60.2.0
six==1.16.0
tensorboard-data-server==0.7.0
tensorboard==2.12.3
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing-extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
wheel==0.37.1
wrapt==1.14.1
